update securityConfig and remove userSettingController
